feat(stream): Decode benchamrk price

+++ src/main/java/org/example/PriceData.java
@@ -1,15 +1,17 @@
 package org.example;
 
+import java.math.BigDecimal;
+
 public class PriceData {
-    private final int price;
+    private final BigDecimal price;
     private final long timestamp;
 
-    private PriceData(long timestamp, int price) {
+    private PriceData(long timestamp, BigDecimal price) {
         this.price = price;
         this.timestamp = timestamp;
     }
 
-    public int getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
@@ -17,7 +19,7 @@ public class PriceData {
         return timestamp;
     }
 
-    public static PriceData fromReport(long timestamp, int price) {
+    public static PriceData fromReport(long timestamp, BigDecimal price) {
         // Parse the report to extract price and timestamp
         // This is just an example; you'll need to implement the actual parsing logic
         //ouble price = 0; // Extracted price
